fix FacetsOrderer tests
